+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 * Model and response classes implement `Serializable` (#57)
 * Split `SercretResponse`  into `PlainSecretResponse` and `MetaSecretResponse` subclasses (common API unchanged)
 * Add missing fields to `AuthMethod` model
+* Add support for (dis)allowed policy glob patterns in `TokenRole`
 
 ### Test
 * Tested against Vault 1.10.1